requirements.txt
mysql-connector-python
tkinter
pandas
numpy
matplotlib
scikit-learn
seaborn
scrapy
scrapy-playwright